+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
 
 Here is a template for new release sections
 
+## [Unreleased]
+### Added
+- clean layer activation
+
+### Changed
+- removed legacy layer handling
+
 ## [0.19.0] - 2024-03-04
 ### Changed
 - remove colorbrewer dependency and implement colors locally