+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v
 
 Here is a template for new release sections
 
-## [Unreleased]
+## [1.1.0] - 2024-03-14
 ### Added
 - error message for missing layer style
 - layers reuse other sources and source layers if model managers match